Move main function into its own package
