+++ get_ocp_range.go
@@ -6,13 +6,9 @@ package getocprange
 import (
 	"embed"
 	"fmt"
-	"os"
 	"strings"
-)
 
-import (
 	"github.com/Masterminds/semver/v3"
-	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3"
 )
 
@@ -139,26 +135,4 @@ func isRangeOpenEnded(kubeVersionRangeConstraint *semver.Constraints) bool {
 	nextUpperKubeVersion := upperKubeVersion.IncMinor()
 	isOpenEnded, _ := kubeVersionRangeConstraint.Validate(&nextUpperKubeVersion)
 	return isOpenEnded
-}
-
-var rootCmd = &cobra.Command{
-	Use:   "get-ocp-version <kubeVersionRange>",
-	Short: "get-ocp-version",
-	Long:  `get-ocp-version derives a range of OCP versions from a given range of Kubernetes Version. It uses Mastermind/semver/v3 under the hood.`,
-	Args:  cobra.ExactArgs(1),
-	RunE: func(cmd *cobra.Command, args []string) error {
-		cmd.SilenceUsage = true
-		resultOCPRange, err := GetOCPRange(args[0])
-		if err != nil {
-			return err
-		}
-		fmt.Println(resultOCPRange)
-		return nil
-	},
-}
-
-func main() {
-	if err := rootCmd.Execute(); err != nil {
-		os.Exit(1)
-	}
 }
